# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "MyBrand-Blog API"
	},
	"basePath": "/api",
	"schemes": [
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
		"/blogs": {
			"get": {
				"summary": "Get all blogs",
				"description": "Returns a list of all blogs",
				"responses": {
					"200": {
						"description": "A list of all blogs",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Blog"
							}
						}
					},
					"500": {
						"description": "Server error"
					}
				}
			},
			"post": {
				"summary": "Create a new blog",
				"description": "Creates a new blog with the given data",
				"parameters": [{
					"name": "blog",
					"in": "body",
					"description": "The data for the new blog",
					"required": true,
					"schema": {
						"$ref": "#/definitions/BlogInput"
					}
				}],
				"responses": {
					"201": {
						"description": "The newly created blog",
						"schema": {
							"$ref": "#/definitions/Blog"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"500": {
						"description": "Server error"
					}
				}
			}
		},
		"/blogs/{id}": {
			"get": {
				"summary": "Get a blog by ID",
				"description": "Returns the blog with the given ID",
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "The ID of the blog",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "The blog with the given ID",
						"schema": {
							"$ref": "#/definitions/Blog"
						}
					},
					"404": {
						"description": "Blog not found"
					},
					"500": {
						"description": "Server error"
					}
				}
			},
			"put": {
				"summary": "Update a blog by ID",
				"description": "Updates the blog with the given ID with the given data",
				"parameters": [{
						"name": "id",
						"in": "path",
						"description": "The ID of the blog",
						"required": true,
						"type": "string"
					},
					{
						"name": "blog",
						"in": "body",
						"description": "The updated data for the blog",
						"required": true,
						"schema": {
							"$ref": "#/definitions/BlogInput"
						}
					}
				],
				"responses": {
					"200": {
						"description": "The updated blog",
						"schema": {
							"$ref": "#/definitions/Blog"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"404": {
						"description": "Blog not found"
					},
					"500": {
						"description": "Server error"
					}
				}
			},
			"delete": {
				"summary": "Delete a blog by ID",
				"description": "Deletes the blog with the given ID",
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "The ID of the blog",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"204": {
						"description": "Blog deleted"
					},
					"404": {
						"description": "Blog not found"
					},
					"500": {
						"description": "Server error"
					}
				}
			}
		}
	},
	"definitions": {
		"Blog": {
			"type": "object",
			"properties": {
				"id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"content": {
					"type": "string"
				},
				"created_at": {
					"type": "string",
					"format": "date-time"
				},
				"updated_at": {
					"type": "string",
					"format": "date-time"
				}
			}
		},
		"BlogInput": {
			"type": "object",
			"properties": {
				"title": {
					"type": "string"
				},
				"content": {
					"type": "string"
				}
			}
		}
	}
}